Safety - if log is empty, don't process

+++ run.sh
@@ -5,6 +5,7 @@ bluetoothctl disconnect DA:AB:E8:9C:B4:31
 date
 ~/.local/bin/meshtastic --ble DA:AB:E8:9C:B4:31 --info >meshtastic.log
 date
-
-python3 script.py
-cp meshtastic_nodes.html /var/www/html/reh.html
+if [ -s meshtastic.log ]; then
+  python3 script.py
+  cp meshtastic_nodes.html /var/www/html/reh.html
+fi
